Hugging_Face_Spaces_Demo/Hugging_Face_Spaces_for_PyThorch/requirements.txt
torch
torchvision
torchaudio
gradio
Pillow
numpy
